Added Methods for Reading and Writing Files to Server
Note - there is a single method you call in any client level class that is marked at the bottom of client. If you just want to be given a file, send an empty string over as contents and then nothing will be added to the files on the server.

Harshini's code is broken so you will need to comment it out.